bug fixes in config

+++ src/main/scala/tools/Config.scala
@@ -29,7 +29,7 @@ object Config {
   lazy val dbpediaLookup = Try(config.getString("dbpedia.lookup")).getOrElse("http://lookup.dbpedia.org/api/search/KeywordSearch?")
 
   lazy val spotlightUrl = Try(config.getString("spotlight.url")).getOrElse("http://spotlight.dbpedia.org")
-  lazy val spotlightConfidence = Try(config.getInt("spotlight.confidence")).getOrElse(0.5)
+  lazy val spotlightConfidence = Try(config.getDouble("spotlight.confidence")).getOrElse(0.5)
   lazy val spotlightSupport = Try(config.getInt("spotlight.support")).getOrElse(10)
 
   lazy val clavinUrl = Try(config.getString("clavin.url")).getOrElse("http://134.169.32.169:9093")
